chore: simplification wave B — adopt @kagan/shared-api-client
Make the previously-orphaned shared package load-bearing as the single
source for cross-package types and shared event-rendering logic.

- packages/shared/api-client/ wired as workspace dep of web and vscode
- packages/shared/api-client/src/event-rendering.ts is now canonical;
  web and vscode each ship a 13-line re-export shim instead of the
  234-line duplicate. Picked the vscode behaviour (AGENT_STATUS → null,
  internal heartbeat) as canonical — silent web improvement
- packages/vscode/src/api/types.ts re-exports EVENT_TYPE, SSE_TYPE,
  TASK_COLUMNS, and the primitive enum types from shared. Adds
  CRITERION_VERDICT and AUTO_REVIEW_STARTED to vscode's event vocab,
  which the local copy was missing
- WireTask kept locally in vscode with strict non-optional fields
  because vscode consumers access them without null checks (server
  always populates them — verified in earlier wave)
- Deleted the duplicated event-rendering.test.ts in vscode; the
  identical web test now exercises the shared implementation via the shim
- packages/shared/api-client/tsconfig switched to CommonJS for vscode's
  Node16 module resolution

Net -268 LOC across the three packages; behaviour neutral except the
AGENT_STATUS rendering improvement on the web side.
