+++ packages/shared/api-client/package.json
@@ -2,12 +2,12 @@
   "name": "@kagan/shared-api-client",
   "version": "0.1.0",
   "description": "Shared TypeScript API client for Kagan - works in both browser and VS Code extension contexts",
-  "type": "module",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "exports": {
     ".": {
-      "import": "./dist/index.js",
+      "require": "./dist/index.js",
+      "default": "./dist/index.js",
       "types": "./dist/index.d.ts"
     }
   },
