+++ packages/vscode/package.json
@@ -403,6 +403,9 @@
     "package": "pnpm run check-types && node esbuild.js --production",
     "vsix": "pnpm exec vsce package --allow-missing-repository"
   },
+  "dependencies": {
+    "@kagan/shared-api-client": "workspace:*"
+  },
   "devDependencies": {
     "@types/mocha": "^10.0.10",
     "@types/node": "^25.5.2",
